Fix Style/FrozenStringLiteralComment
Reason for changing spec/system/validation_spec.rb.

I removed the warning, but I received an error message like this.

```
Failure/Error: text << 'a'

FrozenError:
  can't modify frozen String
```

`<<` method is a destructive method. The `frozen_string_literal: true` will` freeze` all character literals.
If it is `+`, a warning is not generated because it returns a new string.